requirements.txt
ipython
scikit-learn
joblib
pytest
numpy
pandas
matplotlib
dvc
feast
dvc[gs]
black
flake8
mlflow
google-cloud-aiplatform
google-cloud-storage
dvc
#sponsible AI & Monitoring
shap
fairlearn
evidently
sniffio
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
dvc-gs
locust
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-asgi
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions
opentelemetry-util-http
